Revert "Linux4.9.219"

+++ mm/mlock.c
@@ -650,11 +650,11 @@ static int apply_vma_lock_flags(unsigned long start, size_t len,
  * is also counted.
  * Return value: previously mlocked page counts
  */
-static unsigned long count_mm_mlocked_page_nr(struct mm_struct *mm,
+static int count_mm_mlocked_page_nr(struct mm_struct *mm,
 		unsigned long start, size_t len)
 {
 	struct vm_area_struct *vma;
-	unsigned long count = 0;
+	int count = 0;
 
 	if (mm == NULL)
 		mm = current->mm;
